fix: address ANALYSIS.md findings — P0-P6 improvements
P0 (Critical):
- reflect.c: add uv_mutex_t to g_registry for thread safety
- websocket.c: add SIZE_MAX overflow check before malloc(header_len + len)
- server.c: on_body realloc failure frees+nulls body, returns HPE_USER
- server.c: set _internal_client=NULL in on_close/on_timer_close

P1 (Resource management):
- arena.c: check size > SIZE_MAX-7 before (size+7)&~7 alignment
- server.c: check uv_async_init, uv_signal_init, uv_signal_start returns
- utils.c: change csilk_sha1_update len param to size_t

P2 (API design):
- csilk_cors_middleware: pass config by const pointer
- csilk_server_use: return int (-1 when full) instead of silent drop
- csilk_static: add @note about URL prefix stripping

P3 (Code quality):
- advanced_server.c: free groups (api, root) in cleanup
- example_server.c: remove redundant handlers[1]=NULL
- test_ip.c: remove fake pointer (void*)0x1
- test_main.c: remove dead unreferenced file

P4 (New features):
- csilk_redirect(c, status, location): HTTP redirect helper
- csilk_csrf_generate_token(buf, size): random CSRF token generator
- csilk_server_set_max_connections(server, max): concurrent conn limit

P6 (Build system):
- CMake: PRIVATE visibility for yaml/zlib libs
- CI build.yml: add libyaml-dev zlib1g-dev
- CI ci.yml: add zlib1g-dev to both jobs
- CMake: explicit pthread link for test targets

+++ include/csilk_internal.h
@@ -25,7 +25,7 @@ void csilk_sha1_init(csilk_sha1_ctx* context);
  * @param context SHA1 context.
  * @param data Input data bytes.
  * @param len Length of input data. */
-void csilk_sha1_update(csilk_sha1_ctx* context, const uint8_t* data, uint32_t len);
+void csilk_sha1_update(csilk_sha1_ctx* context, const uint8_t* data, size_t len);
 
 /** @brief Finalize SHA1 hash and produce the digest.
  * @param context SHA1 context.
